move non-assertion-functions from assert to test and avoid duplicate code
