+++ test/metadata.go
@@ -1,4 +1,4 @@
-package assert
+package test
 
 import (
 	"bufio"
